Adjust surface variant decorator.
